// frontend/src/app/model/Person.ts
import { User } from './User';

export type Person = {
  id: string;
  user: User;
  user_id: string;
  image: string;
  name: string;
  cpf: string;
  birth_date: string;
  email: string;
  phone_one: string;
  phone_two: string;
  sex: string;
  cep: string;
  street: string;
  number: string;
  complement: string;
  district: string;
  city: string;
  state: string;
  country: string;
  created_at: string;
  updated_at: string;
};